Handle bundled modules.
Make several changes to handle bundling modules inside of a package's
node_modules directory.

Do not raise an error if node_modules already exists. If a dependency
name already exists as node_modules/<depName>, assume it is correct and
continue.

When linking a module that is bundled within another module, look for
the dependencies both within the node_modules directory of the outer
module, and as the outer module itself. This allows for bundling to
break dependency cycles.